more bug fixes + flameburst / breath animations

+++ src/com/ebicep/warlords/classes/AbstractAbility.java
@@ -14,13 +14,13 @@ public abstract class AbstractAbility {
     protected int minDamageHeal;
     protected int maxDamageHeal;
     protected float currentCooldown;
-    protected int cooldown;
+    protected float cooldown;
     protected int energyCost;
     protected int critChance;
     protected int critMultiplier;
     protected String description;
 
-    public AbstractAbility(String name, int minDamageHeal, int maxDamageHeal, int cooldown, int energyCost, int critChance, int critMultiplier, String description) {
+    public AbstractAbility(String name, int minDamageHeal, int maxDamageHeal, float cooldown, int energyCost, int critChance, int critMultiplier, String description) {
         this.name = name;
         this.minDamageHeal = minDamageHeal;
         this.maxDamageHeal = maxDamageHeal;
@@ -71,11 +71,11 @@ public abstract class AbstractAbility {
         }
     }
 
-    public int getCooldown() {
+    public float getCooldown() {
         return cooldown;
     }
 
-    public void setCooldown(int cooldown) {
+    public void setCooldown(float cooldown) {
         this.cooldown = cooldown;
     }
